Remove type conversion for hour column since it gets dropped

+++ poll.py
@@ -105,11 +105,10 @@ def tabularise(payload: dict[str, Any]) -> pd.DataFrame:
     result.columns = pd.Index(new_cols)
 
     # Step 8: Apply type conversions (matching Power Query)
-    result["hour"] = result["hour"].astype("int64")
+    # Note: hour, venue_id, and Attribute are not converted since they will be dropped in Step 10
     result["fromTime"] = pd.to_datetime(result["fromTime"], format="%H:%M", errors="coerce").dt.time
     result["day"] = pd.to_datetime(result["day"], errors="coerce").dt.date
     result["Value.total_spaces"] = result["Value.total_spaces"].astype("int64")
-    # Note: venue_id is not converted here since it will be dropped in Step 10
     result["Value.spaces.total_spaces"] = result["Value.spaces.total_spaces"].astype("int64")
     result["scraped_at"] = pd.to_datetime(result["scraped_at"], errors="coerce")
